+++ include/coap2/coap_internal.h
@@ -50,9 +50,11 @@
 #include "coap_async_internal.h"
 #include "coap_block_internal.h"
 #include "coap_cache_internal.h"
+#include "coap_io_internal.h"
 #include "coap_net_internal.h"
 #include "coap_session_internal.h"
 #include "coap_resource_internal.h"
+#include "coap_session_internal.h"
 #include "coap_subscribe_internal.h"
 #include "coap_tcp_internal.h"
