+++ include/coap2/coap_session_internal.h
@@ -16,6 +16,8 @@
 #ifndef COAP_SESSION_INTERNAL_H_
 #define COAP_SESSION_INTERNAL_H_
 
+#include "coap_io_internal.h"
+
 #define COAP_DEFAULT_SESSION_TIMEOUT 300
 #define COAP_PARTIAL_SESSION_TIMEOUT_TICKS (30 * COAP_TICKS_PER_SECOND)
 #define COAP_DEFAULT_MAX_HANDSHAKE_SESSIONS 100
